fix: call api to api service
