+++ src/server/api/product/Product.get.ts
@@ -1,46 +1,12 @@
 import {Product} from "../../../models/Product";
+import ApiService from "../ApiService";
 
 export const getProductById = async (productId: number): Promise<Product> => {
-    try {
-        const token = localStorage.getItem("authToken");
-        const response = await fetch(`/api/product/${productId}`, {
-            method: "GET",
-            headers: {
-                "Content-Type": "application/json",
-                Authorization: `Bearer ${token}`,
-            }
-        })
-        if (!response.ok) {
-            const data = await response.json();
-            throw new Error(data.message || "Unauthorized or error in response");
-        }
-        const data = await response.json();
-        return data.data;
-
-    } catch (err) {
-        console.error("Error:", err);
-        throw new Error("Failed to fetch product");
-    }
+    const {data} = await ApiService.get(`/api/product/${productId}`);
+    return data;
 }
 export const getListProduct = async (): Promise<Product[]> => {
-    try {
-        const token = localStorage.getItem("authToken");
-        const response = await fetch("/api/product/list", {
-            method: "GET",
-            headers: {
-                "Content-Type": "application/json",
-                Authorization: `Bearer ${token}`,
-            },
-        });
-        if (!response.ok) {
-            const data = await response.json();
-            throw new Error(data.message || "Unauthorized or error in response");
+    const {data} = await ApiService.get("/api/product/list");
+    return data;
 
-        }
-        const data = await response.json();
-        return data.data;
-    } catch (err) {
-        console.error("Error:", err);
-        throw new Error("Failed to fetch product");
-    }
 }
